internal: move doc_test mod into moonutil
