+++ crates/moon/Cargo.toml
@@ -51,7 +51,6 @@ tokio.workspace = true
 futures.workspace = true
 clap_complete.workspace = true
 indexmap.workspace = true
-regex.workspace = true
 
 [target.'cfg(not(windows))'.dependencies]
 openssl = { version = "0.10.66", features = ["vendored"] }
