+++ crates/moonutil/Cargo.toml
@@ -47,6 +47,7 @@ schemars.workspace = true
 line-index.workspace = true
 nucleo-matcher.workspace = true
 which.workspace = true
+regex.workspace = true
 
 [dev-dependencies]
 expect-test.workspace = true
