+++ crates/moonutil/src/lib.rs
@@ -35,3 +35,4 @@ pub mod path;
 pub mod render;
 pub mod scan;
 pub mod version;
+pub mod doc_test;
